+++ dev/server/graphql/src/main.rs
@@ -1,26 +1,6 @@
-use async_graphql::{http::GraphiQLSource, Schema};
-use async_graphql_axum::{GraphQL, GraphQLSubscription};
-use axum::{
-    response::{self, IntoResponse},
-    routing::get,
-    Router,
-};
-use relentless_graphql_dev_server::{MutationRoot, QueryRoot, Storage, SubscriptionRoot};
-use tokio::net::TcpListener;
-
-async fn graphiql() -> impl IntoResponse {
-    response::Html(GraphiQLSource::build().endpoint("/").subscription_endpoint("/ws").finish())
-}
+use relentless_graphql_dev_server::serve;
 
 #[tokio::main]
-async fn main() {
-    let schema = Schema::build(QueryRoot, MutationRoot, SubscriptionRoot).data(Storage::default()).finish();
-
-    let app = Router::new()
-        .route("/", get(graphiql).post_service(GraphQL::new(schema.clone())))
-        .route_service("/ws", GraphQLSubscription::new(schema));
-
-    println!("GraphiQL IDE: http://localhost:8000");
-
-    axum::serve(TcpListener::bind("127.0.0.1:8000").await.unwrap(), app).await.unwrap();
+async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    serve().await
 }
